revert to 32-bit objd values
The size of the objd->value member was previously increased to 64-bit
to allow setting 64-bit default values. This may waste some (ROM)
memory as 64-bit values are somewhat uncommon, and the value can be
set from an application hook instead. Revert to 32-bit values.

Fix #59

+++ soes/esc_coe.h
@@ -21,7 +21,7 @@ typedef struct CC_PACKED
    uint16_t bitlength;
    uint16_t flags;
    const char *name;
-   uint64_t value;
+   uint32_t value;
    void *data;
 } _objd;
 CC_PACKED_END
